neorpc/result: move NotaryRequestEvent to this package
Not worth a package of its own.
